+++ lib/src/model/inheritable.dart
@@ -16,13 +16,33 @@ import 'package:dartdoc/src/special_elements.dart';
 /// namespace, that's the one we should treat as canonical and implementors
 /// of this class can use that knowledge to determine canonicalization.
 ///
-/// We pick the class closest to the [definingEnclosingElement] so that all
+/// We pick the class closest to the [definingEnclosingContainer] so that all
 /// children of that class inheriting the same member will point to the same
 /// place in the documentation, and we pick a canonical class because that's
 /// the one in the public namespace that will be documented.
 mixin Inheritable on ContainerMember {
+  /// True if this [Inheritable] is inherited from a different class.
+  bool get isInherited;
+
+  /// True if this [Inheritable] has a parameter whose type is overridden
+  /// by a subtype.
+  bool get isCovariant;
+
+  @override
+  Set<String> get features {
+    Set<String> _features = super.features;
+    if (isOverride) _features.add('override');
+    if (isInherited) _features.add('inherited');
+    if (isCovariant) _features.add('covariant');
+    return _features;
+  }
+
+  @override
+  Library get canonicalLibrary => canonicalEnclosingContainer?.canonicalLibrary;
+
   @override
   ModelElement buildCanonicalModelElement() {
+    // TODO(jcollins-g): factor out extension logic into [Extendable]
     if (canonicalEnclosingContainer is Extension) {
       return this;
     }
@@ -33,6 +53,9 @@ mixin Inheritable on ContainerMember {
               (m) =>
                   m.name == name && m.isPropertyAccessor == isPropertyAccessor,
               orElse: () => null);
+    }
+    if (canonicalEnclosingContainer != null) {
+      throw UnimplementedError('${canonicalEnclosingContainer}: unknown type');
     }
     return null;
   }
@@ -77,8 +100,14 @@ mixin Inheritable on ContainerMember {
       if (definingEnclosingContainer.isCanonical &&
           definingEnclosingContainer.isPublic) {
         assert(definingEnclosingContainer == found);
+      }
+      if (found != null) {
         return found;
       }
+    } else if (!isInherited && definingEnclosingContainer is! Extension) {
+      // TODO(jcollins-g): factor out extension logic into [Extendable].
+      return packageGraph
+          .findCanonicalModelElementFor(element.enclosingElement);
     }
     return super.computeCanonicalEnclosingContainer();
   }
@@ -104,7 +133,7 @@ mixin Inheritable on ContainerMember {
 
   bool _isOverride;
 
-  @override
+  /// True if this [Inheritable] is overriding a superclass.
   bool get isOverride {
     if (_isOverride == null) {
       // The canonical version of the enclosing element -- not canonicalEnclosingElement,