+++ lib/src/model/model_element.dart
@@ -202,7 +202,8 @@ abstract class ModelElement extends Canonicalization
         e is ParameterElement ||
         e is TypeParameterElement ||
         e is GenericFunctionTypeElementImpl ||
-        e.kind == ElementKind.DYNAMIC);
+        e.kind == ElementKind.DYNAMIC ||
+        e.kind == ElementKind.NEVER);
 
     Member originalMember;
     // TODO(jcollins-g): Refactor object model to instantiate 'ModelMembers'
@@ -215,12 +216,16 @@ abstract class ModelElement extends Canonicalization
         Tuple3(e, library, enclosingContainer);
     ModelElement newModelElement;
     if (e.kind != ElementKind.DYNAMIC &&
+        e.kind != ElementKind.NEVER &&
         packageGraph.allConstructedModelElements.containsKey(key)) {
       newModelElement = packageGraph.allConstructedModelElements[key];
       assert(newModelElement.element is! MultiplyInheritedExecutableElement);
     } else {
       if (e.kind == ElementKind.DYNAMIC) {
         newModelElement = Dynamic(e, packageGraph);
+      }
+      if (e.kind == ElementKind.NEVER) {
+        newModelElement = NeverType(e, packageGraph);
       }
       if (e is MultiplyInheritedExecutableElement) {
         newModelElement = resolveMultiplyInheritedElement(
@@ -642,7 +647,7 @@ abstract class ModelElement extends Canonicalization
     if (!_canonicalLibraryIsSet) {
       // This is not accurate if we are constructing the Package.
       assert(packageGraph.allLibrariesAdded);
-      // Since we're may be looking for a library, find the [Element] immediately
+      // Since we're looking for a library, find the [Element] immediately
       // contained by a [CompilationUnitElement] in the tree.
       Element topLevelElement = element;
       while (topLevelElement != null &&
@@ -739,6 +744,7 @@ abstract class ModelElement extends Canonicalization
 
   @override
   bool get isCanonical {
+    if (!isPublic) return false;
     if (library == canonicalLibrary) {
       if (this is Inheritable) {
         Inheritable i = (this as Inheritable);
@@ -1120,6 +1126,7 @@ abstract class ModelElement extends Canonicalization
     // element associated with a ModelElement or there's an analysis bug.
     assert(name.isNotEmpty ||
         this.element?.kind == ElementKind.DYNAMIC ||
+        this.element?.kind == ElementKind.NEVER ||
         this is ModelFunction);
 
     if (href == null) {