+++ lib/src/warnings.dart
@@ -206,6 +206,19 @@ final Map<PackageWarning, PackageWarningDefinition> packageWarningDefinitions =
         "documentation comments to hide it."
       ],
       defaultWarningMode: PackageWarningMode.error),
+  PackageWarning.missingConstantConstructor: PackageWarningDefinition(
+      PackageWarning.missingConstantConstructor,
+      "missing-constant-constructor",
+      "Dartdoc can not show the value of a constant because its constructor could not be resolved.",
+      longHelp: [
+        "To resolve a constant into its literal value, Dartdoc relies on the",
+        "analyzer to resolve the constructor.  The analyzer didn't provide",
+        "the constructor for @@name@@, which is usually due to an error in the",
+        "code.  Use the analyzer to find missing imports.",
+      ],
+      // Defaults to ignore as this doesn't impact the docs severely but is
+      // useful for debugging package structure.
+      defaultWarningMode: PackageWarningMode.ignore),
 };
 
 /// Something that package warnings can be called on.  Optionally associated
@@ -244,6 +257,7 @@ enum PackageWarning {
   toolError,
   deprecated,
   unresolvedExport,
+  missingConstantConstructor,
 }
 
 /// Used to declare defaults for a particular package warning.